First checkin of an AppleScript that dumps out data from Palm Desktop.  It uses a custom XML format to represent the following:   Categories   Addresses   Memos   Events   To Dos It attempts to preserve "attachments" between items, such as Addresses attached to events, memos attached to memos, etc.
Currently, this script runs exceedingly slowly.
Also, it probably needs a copy of SKProgressBar to even compile.
The progress bar is currently disabled, as it was incorrectly suspected as
being the cause of the speed problem.
The IDs for representing the links between items are not done correctly, as
the ID assigned internally to each item is unique per type of item, not for all
items.